refactor: extract VPN auth logic
